admin-dashboard-final-test

+++ backend/backend/services/projectService.js
@@ -384,9 +384,12 @@ module.exports = {
         const data = { stripePlanId: 'enterprise', stripeSubscriptionId: null };
         try {
             const project = await this.findOneBy({ _id: projectId });
-            await PaymentService.removeSubscription(
-                project.stripeSubscriptionId
-            );
+            console.log("Project 1: ",data)
+            if (data.stripeSubscriptionId !== null){
+                await PaymentService.removeSubscription(
+                    project.stripeSubscriptionId
+                );
+            }
             const updatedProject = await this.updateOneBy(
                 { _id: projectId },
                 data
